text2image/text2image_service_requirements.txt
langchain_community
langchain_openai
streamlit
langchain_core
python-dotenv
langserve
fastapi
uvicorn
ipykernel
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
arxiv
wikipedia
langchainhub
langchain-groq
groq